finish show meta

+++ src/main/java/cn/edu/thssdb/parser/ImpVisitor.java
@@ -16,6 +16,7 @@ import javax.management.AttributeNotFoundException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
+import java.util.StringJoiner;
 
 /**
  * When use SQL sentence, e.g., "SELECT avg(A) FROM TableX;"
@@ -56,6 +57,7 @@ public class ImpVisitor extends SQLBaseVisitor<Object> {
         if (ctx.use_db_stmt() != null)  return new QueryResult(visitUse_db_stmt(ctx.use_db_stmt()));
         if (ctx.create_table_stmt() != null) return new QueryResult(visitCreate_table_stmt(ctx.create_table_stmt()));
         if (ctx.drop_table_stmt() != null) return new QueryResult(visitDrop_table_stmt(ctx.drop_table_stmt()));
+        if (ctx.show_meta_stmt() != null) return new QueryResult(visitShow_meta_stmt(ctx.show_meta_stmt()));
         if (ctx.insert_stmt() != null) return new QueryResult(visitInsert_stmt(ctx.insert_stmt()));
         if (ctx.delete_stmt() != null) return new QueryResult(visitDelete_stmt(ctx.delete_stmt()));
         if (ctx.update_stmt() != null) return new QueryResult(visitUpdate_stmt(ctx.update_stmt()));
@@ -251,6 +253,16 @@ public class ImpVisitor extends SQLBaseVisitor<Object> {
             }
         }
         return "Inserted " + ctx.value_entry().size() + " rows.";
+    }
+
+    @Override
+    public String visitShow_meta_stmt(SQLParser.Show_meta_stmtContext ctx) {
+        try {
+            return GetCurrentDB().getTableInfo(ctx.table_name().getText());
+        }
+        catch (Exception e) {
+            return e.getMessage();
+        }
     }
 
     public String[] visitValue_entry(SQLParser.Value_entryContext context) {
